Feat: move new field form from a dialog to a page
